separated endpoints non-dependent by token and replaced query params by request body for security reasons
