+++ src/main/java/dlt/dltbackendmaster/controller/UserController.java
@@ -6,7 +6,6 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.repository.query.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -21,9 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 import dlt.dltbackendmaster.domain.Users;
 import dlt.dltbackendmaster.security.EmailSender;
 import dlt.dltbackendmaster.security.utils.PasswordGenerator;
-import dlt.dltbackendmaster.security.utils.Utility;
 import dlt.dltbackendmaster.service.DAOService;
-import net.bytebuddy.utility.RandomString;
 
 @RestController
 @RequestMapping("/api/users")
@@ -112,49 +109,18 @@ public class UserController
         }
     }
 
-    @PutMapping(path = "/update-password", produces = "application/json")
-    public ResponseEntity<Users> updatePassword(HttpServletRequest request, @Param(value = "username") String username, @Param(value = "newPassword") String newPassword) {
-
-        Users user = service.GetUniqueEntityByNamedQuery("Users.findByUsername", username);
-
-        if (username == null || user == null) {
-            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
-        }
-
-        try {
-            user.setRecoverPassword(passwordEncoder.encode(newPassword));
-            user.setNewPassword(0);
-            user.setIsEnabled(Byte.valueOf("0"));
-            String token = RandomString.make(45);
-            user.setRecoverPasswordToken(token);
-            Users updatedUser = service.update(user);
-            // Generate reset confirmation Link
-            String confirmUpdatePasswordLink = Utility.getSiteURL(request) + "/confirm-update?token=" + token;
-            // Send E-mail
-            String email = user.getEmail() != null ? user.getEmail() : user.getUsers().getEmail();
-            emailSender.sendEmail(user.getName() + " " + user.getSurname(),
-                                  null,
-                                  email,
-                                  confirmUpdatePasswordLink,
-                                  false);
-            return new ResponseEntity<>(updatedUser, HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
     @PutMapping(path = "/change-password", produces = "application/json")
-    public ResponseEntity<Users> changePassword(HttpServletRequest request, @Param(value = "username") String username, @Param(value = "newPassword") String newPassword) {
+    public ResponseEntity<Users> changePassword(HttpServletRequest request, @RequestBody Users users) {
   
-        Users user = service.GetUniqueEntityByNamedQuery("Users.findByUsername", username);
+        Users user = service.GetUniqueEntityByNamedQuery("Users.findByUsername", users.getUsername());
 
-        if (username == null || user == null) {
+        if (user == null) {
             return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
         }
 
         try {
             user.setNewPassword(0);
-            user.setPassword(passwordEncoder.encode(newPassword));
+            user.setPassword(passwordEncoder.encode(users.getRecoverPassword()));
             Users updatedUser = service.update(user);
             
             return new ResponseEntity<>(updatedUser, HttpStatus.OK);
