+++ src/main/java/dlt/dltbackendmaster/security/SecurityConfigurations.java
@@ -77,8 +77,8 @@ public class SecurityConfigurations extends WebSecurityConfigurerAdapter{
             .antMatchers(HttpMethod.GET, "/sync").permitAll()
             .antMatchers(HttpMethod.OPTIONS, "/api/login").permitAll()
             //allow anonymous to Confirm Update Password
-            .antMatchers(HttpMethod.PUT, "/confirm-update").permitAll()
-            .antMatchers(HttpMethod.OPTIONS, "/confirm-update").permitAll()
+            .antMatchers(HttpMethod.PUT, "/users").permitAll()
+            .antMatchers(HttpMethod.OPTIONS, "/users").permitAll()
             	//allow anonymous POSTs to login
             .antMatchers(HttpMethod.POST, "/api/login").permitAll()
             .antMatchers(HttpMethod.OPTIONS, "/api/login").permitAll()
